requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
docopt==0.6.2
itsdangerous==0.24
joblib==0.8.4
# julia==0.1.1
numpy==1.9.0
pandas==0.14.1
psutil==2.1.3
py==1.4.26
python-dateutil==2.4.1
pytz==2014.7
# rpy2==2.5.2
scikit-learn==0.15.2
scipy==0.14.0
six==1.9.0
sqlalchemy==0.9.8
ujson==1.33
